src/requirements.txt
numba==0.47.0
pyteomics==4.1.1
numpy==1.18.1
more_itertools==7.1.0
pandas==0.25.3
faiss-gpu==1.5.3
dleamse==0.2.7
click
scipy
h5py
